moved msg_register_host_zone
